Заменил двумерные double массивы на одномерные int.

+++ src/Application/DTO/ProcessingRequest.cs
@@ -6,6 +6,6 @@ public class ProcessingRequest
     public Stream OriginalStream { get; set; } = null!;
     public int CountPoints { get; set; } = 240;
     public int CountSteps { get; set; } = 4000;
-    public double ContrastLine { get; set; } = 15;
+    public int ContrastLine { get; set; } = 15;
     public int Padding { get; set; } = 10;
 }
